Initial commit of TMD schema in JSON-Schema format
This initial commit contains the current data schema specification for Target Market Determinations as of the 2021-04-26 version of FSC DDO Data Standards documentation.

This schema is auto-generated by the data interface platform at APIR Systems and is a work in progress.  Notable admissions are enumerations which still need to be ported from external source formats to JSON-Schema.  But for the sake of "getting something out there" what we have is being committed now.